requirements.txt
fastapi
uvicorn
gunicorn
google-cloud-vision
python-multipart